surya-ocr-demo/pyproject.toml
[project]
name = "surya-ocr-demo"
version = "0.1.0"
description = "Marimo notebook demo for SuryaOCR"
requires-python = ">=3.10"
authors = [
    {name = "ahmed-0011"},
]
dependencies = [
    "pyperclip>=1.9.0",
    "marimo[recommended]>=0.14.11,<1.0",
    "opencv-python~=4.12.0.88",
    "setuptools",
    "torch>=2.8.0",
    "surya-ocr>=0.15.2",
    "torchvision>=0.23.0",
]

[tool.marimo.package_management]
manager = "uv"

[tool.marimo.runtime]
auto_instantiate = true
on_cell_change = "autorun"
